PRSD-362: Refactor OS Places stub for better organisation
