+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/local/api/controller/OSPlacesAPIStubController.kt
@@ -1,12 +1,14 @@
-package uk.gov.communities.prsdb.webapp.local.api
+package uk.gov.communities.prsdb.webapp.local.api.controller
 
 import org.springframework.context.annotation.Profile
 import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
 
 @Profile("local-mock-os-places")
 @RestController
+@RequestMapping("/os-places-local")
 class OSPlacesAPIStubController {
     @GetMapping("/postcode")
     fun lookupAddressesByPostcode(
